requirements.txt
torch
accelerate
transformers
datasets
jupyterlab
pyyaml
pandas
tqdm
python-dotenv
ipywidgets
pillow